pyproject.toml
[project]
name = "project-2048"
version = "0.1.0"
description = "find interpretable strategy in 2048 game"
authors = [
    {name = "Polina Volodchenko",email = "volodchenko_02@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy (>=2.2.2,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "gym (>=0.26.2,<0.27.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "gymnasium-2048 (>=0.0.2,<0.0.3)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
